<h1> API / Function References </h1> <h2>Calculating Model Input Features</h2>

RadarQC.process_single_file — Function

Driver function that calculates a set of features from a single CFRadial file. Features are 
specified in file located at argfile_path. 

IMPLICITLY specifies weight matrixes (already in calc function args) 

If the file has already been manually QCed, set
HAS_MANUAL_QC to true, and an additional Y array will be returned containing "1" if a datapoint
was retained in manual QC and "0" if it was removed. 
NOTE ARGUMENTS remove_variable and QC_variable . The gates retained after manual QC will be obtained 
by determining which are present in remove_variable but are NOT PRESENT in QC_variable 
    
Also included are flags to remove datapoints 
with low Normalized Coherent Power (NCP) and high Probability of Ground (PGG). Finally, the 
remove_variable argument specifies which field will be used to determine where "missing" data exists 
and will subsequently be removed from the returned dataset. 

Returns: 
    -X: Array that is dimensioned (num_gates x num_features) where num_gates is the number of valid 
        (non-missing, meeting NCP/PGG thresholds) the function finds, and num_features is the 
        number of features specified in the argument file to calculate. 

    -Y: IF HAS_MANUAL_QC == true, will return Y, array containing 1 if a datapoint was retained 
        during manual QC, and 0 otherwise. 

    -INDEXER: Based on remove_variable as described above, contains boolean array specifiying
              where in the scan features valid data and where does not.

RadarQC.QC_scan — Function

Primary function to apply a trained RF model to a cfradial scan

RadarQC.get_task_params — Function

Function to parse a given task list Also performs checks to ensure that the specified tasks are able to be performed to the specified CFRad file

RadarQC.calculate_features — Method

Function to process a set of cfradial files and produce a set of input features for training/evaluating a model inputloc: cfradial files are specified by inputloc - can be either a file or a directory tasks: Features to be calculated weightmatrixes: Vector of matrixes that specify size and weights for each spatial variable outputfile : Location to output the resulting dataset

Keyword Arguments: 
remove_variable - variable in CFRadial file used to determine where 'missing' gates are. 
                  these gates will be removed from the outputted features so that the model 
                  is not trained on missing data

RadarQC.calculate_features — Method

Function to process a set of cfradial files and produce a set of input features for training/evaluating a model inputloc: cfradial files are specified by inputloc - can be either a file or a directory argumetfile: Features to be calculated are specified in a file located at argfile output_file : Location to output the resulting dataset

Keyword Arguments: 
remove_variable - variable in CFRadial file used to determine where 'missing' gates are. 
                  these gates will be removed from the outputted features so that the model 
                  is not trained on missing data

RadarQC.split_training_testing! — Function

Function to split a given directory or set of directories into training and testing files using the configuration described in DesRosiers and Bell 2023

Arguments: DIRPATHS: Vector{String} consisting of directory or list of directories containing input CFRadial files TRAININGPATH: String - path to directory where training files will be linked TESTING_PATH: String - path to directory where testing files will be linked

Configuration info: 80/20 Training/Testing split, with testing cases taken from the beginning, middle, and end of EACH case in DIR_PATHS. The function calculates a number of testing scans and divides it up amongst the cases, so each case will be equally represented in the testing dataset.

In the example given in the paper, there are 4 cases, and an 80/20 split gives a total number of testing scans 
of 356. Divided by 4, this results in 89 scans from each case. 

Details about rounding are contained within the function

IMPORTANT: ASSUMES SCANS FOLLOW CFRAD NAMING CONVENTIONS AND ARE THEREFORE LISTED CHRONOLOGICALLY