Basic Workflow Walkthrough

RadarQC is a package that utilizes a methodology developed by Dr. Alex DesRosiers and Dr. Michael Bell to remove non-meteorological gates from Doppler radar scans by leveraging machine learning techniques. In its current form, it contains functionality to derive a set of features from input radar data, use these features to train a Random Forest classification model, and apply this model to the raw fields contained within the radar scans. It also has some model evaluation ability. The beginning of this guide will walk through a basic workflow to train a model starting from scratch.

API / Function References

Calculating Model Input Features

Ronin.calculate_features — Method

Function to process a set of cfradial files and produce input features for training/evaluating a model

Required arguments

input_loc::String

Path to input cfradial or directory of input cfradials

argument_file::String

Path to configuration file containing which features to calculate

output_file::String

Path to output calculated features to (generally ends in .h5)

HAS_INTERACTIVE_QC::Bool

Specifies whether or not the file(s) have already undergone a interactive QC procedure. If true, function will also output a Y array used to verify where interactive QC removed gates. This array is formed by considering where gates with non-missing data in raw scans (specified by remove_variable) are set to missing after QC is performed.

Optional keyword arguments

verbose::Bool=false

If true, will print out timing information as each file is processed

REMOVE_LOW_NCP::Bool=false

If true, will ignore gates with Normalized Coherent Power/Signal Quality Index below a threshold specified in RQCFeatures.jl

REMOVE_HIGH_PGG::Bool=false

If true, will ignore gates with Probability of Ground Gate (PGG) values at or above a threshold specified in RQCFeatures.jl

QC_variable::String="VG"

Name of variable in input NetCDF files that has been quality-controlled.

remove_variable::String="VV"

Name of a raw variable in input NetCDF files. Used to determine where missing data exists in the input sweeps. Data at these locations will be removed from the outputted features.

replace_missing::Bool=false

Whether or not to replace MISSING values with FILL_VAL in spatial parameter calculations Default value: False

write_out::Bool=true

Whether or not to write features out to file

return_idxer::Bool = false

If true, will return IDXER, where IDXER is a

Ronin.calculate_features — Method

Function to process a set of cfradial files and produce input features for training/evaluating a model. Allows for user-specified tasks and weight matrices, otherwise the same as above.

Required arguments

input_loc::String

Path to input cfradial or directory of input cfradials

tasks::Vector{String}

Vector containing the features to be calculated for each cfradial. Example [DBZ, ISO(DBZ)]

weight_matrixes::Vector{Matrix{Union{Missing, Float64}}}

For each task, a weight matrix specifying how much each gate in a spatial calculation will be given. Required to be the same size as tasks

output_file::String 

Location to output the calculated feature data to.

HAS_INTERACTIVE_QC::Bool

Specifies whether or not the file(s) have already undergone a interactive QC procedure. If true, function will also output a Y array used to verify where interactive QC removed gates. This array is formed by considering where gates with non-missing data in raw scans (specified by remove_variable) are set to missing after QC is performed.

Optional keyword arguments

verbose::Bool = false 

If true, will print out timing information as each file is processed

REMOVE_LOW_NCP::Bool = false 

If true, will ignore gates with Normalized Coherent Power/Signal Quality Index below a threshold specified in RQCFeatures.jl

REMOVE_HIGH_PGG::Bool = false

If true, will ignore gates with Probability of Ground Gate (PGG) values at or above a threshold specified in RQCFeatures.jl

QC_variable::String = "VG"

Name of variable in input NetCDF files that has been quality-controlled.

remove_variable::String = "VV"

Name of a raw variable in input NetCDF files. Used to determine where missing data exists in the input sweeps. 
Data at these locations will be removed from the outputted features. 

replace_missing::Bool = false

Whether or not to replace MISSING values with FILL_VAL in spatial parameter calculations
Default value: False 

julia write_out::Bool = true ``` Whether or not to write out to file.

Ronin.split_training_testing! — Function

Function to split a given directory or set of directories into training and testing files using the configuration described in DesRosiers and Bell 2023. This function assumes that input directories only contain cfradial files that follow standard naming conventions, and are thus implicitly chronologically ordered. The function operates by first dividing file names into training and testing sets following an 80/20 training/testing split, and subsequently softlinking each file to the training and testing directories. Attempts to avoid temporal autocorrelation while maximizing variance by dividing each case into several different training/testing sections.

Required Arguments:

DIR_PATHS::Vector{String}

List of directories containing cfradials to be used for model training/testing. Useful if input data is split into several different cases.

TRAINING_PATH::String 

Directory to softlink files designated for training into.

TESTING_PATH::String 

Directory to softlink files designated for testing into.

Ronin.train_model — Method

Function to train a random forest model using a precalculated set of input and output features (usually output from calculate_features). Returns nothing.

Required arguments

input_h5::String

Location of input features/targets. Input features are expected to have the name "X", and targets the name "Y". This should be taken care of automatically if they are outputs from calculate_features

model_location::String 

Path to save the trained model out to. Typically should end in .jld2

Optional keyword arguments

verify::Bool = false 

Whether or not to output a separate .h5 file containing the trained models predictions on the training set (Y_PREDICTED) as well as the targets for the training set (Y_ACTUAL)

verify_out::String="model_verification.h5"

If verify, the location to output this verification to.

col_subset=: 

Set of columns from input_h5 to train model on. Useful if one wishes to train a model while excluding some features from a training set.

row_subset=:

Set of rows from input_h5 to train on.

n_trees::Int = 21

Number of trees in the Random Forest ensemble

max_depth::Int = 14

Maximum node depth in each tree in RF ensemble

class_weights::Vector{Float32} = Vector{Float32}([1.,2.])

Vector of class weights to apply to each observation. Should be 1 observation per sample in the input data files

Ronin.remove_validation — Function

Function used to remove a given subset of the rows from a feature set so that they may be used for model validation/tuning.

Currently configured to utilize the 90/10 split described in DesRosiers and Bell 2023.

Required arguments

input_dataset::String

Path to h5 files containing model features

Optional keyword arguments

training_output::String = "train_no_validation_set.h5"

Path to output training features with validation removed to

validation_output::String = "validation.h5"

Path to output validation features to

remove_original::Bool = true 

Whether or not to remove the original file described by the input_dataset path.

Ronin.get_feature_importance — Method

Uses L1 regression with a variety of λ penalty values to determine the most useful features for

input to the random forest model.

Required Input

input_file_path::String 

Path to .h5 file containing model training features under ["X"] parameter, and model targets under ["Y"] parameter. Also expects the h5 file to contain an attribute known as Parameters containing abbreviations for the feature types

λs::Vector{Float64}

Vector of values used to vary the strength of the penalty term in the regularization.

Optional Keyword Arguments

pred_threshold::Float64

Minimum cofidence level for binary classifier when predicting

Returns

Returns a DataFrame with each row containing info about a regression for a specific λ, the values of the regression coefficients for each input feature, and the Root Mean Square Error of the resultant regression.

Applying and evaluating a trained model to data

Ronin.QC_scan — Function

Primary function to apply a trained RF model to certain raw fields of a cfradial scan. Values determined to be non-meteorological by the RF model will be replaced with Missing

Required Arguments

file_path::String 

Location of input cfradial or directory of cfradials one wishes to apply QC to

config_file_path::String 

Location of config file containing features to calculate as inputs to RF model

model_path::String 

Location of trained RF model (in jld2 file format)

Optional Arguments

VARIABLES_TO_QC::Vector{String} = ["ZZ", "VV"]

List containing names of raw variables in the CFRadial to apply QC algorithm to.

QC_suffix::String = "_QC"

Used for naming the QC-ed variables in the modified CFRadial file. Field name will be QCsuffix appended to the raw field. Example: `DBZQC`

indexer_var::String = "VV"

Variable used to determine what gates are considered "missing" in the raw moments. QC will not be applied to these gates, they will simply remain missing.

decision_threshold::Float64 = .5

Used to leverage probablistic nature of random forest methodology. When the model has a greater than decision_threshold level confidence that a gate is meteorological data, it will be assigned as such. Anything at or below this confidence threshold will be assigned non-meteorological. At least in the ELDORA case, aggressive thresholds (.8 and above) have been found to maintain

92% of the meteorological data while removing >99% of non-meteorological gates.

output_mask::Bool = true

Whether or not to output the QC preditions from the model output. A value of 0 means the model predicted the gate to be non-meteorological, 1 corresponds to predicted meteorological data, and -1 denotes data that did not meet minimum thresholds

mask_name::String = "QC_MASK"

What to name the output QC predictions.

verbose::Bool = false 
````
Whether to output timing and scan information

julia output_probs::Bool = false

Whether or not to output probabilities of meteorological gate from random forest 

julia prob_varname::String = "" ``` What to name the probability variable in the cfradial file

Version of QC_Scan used for multiple-pass models. Required argument is a model configuration object. Optional arguments:

output_probs::Bool = false 

Whether or not to write the modeled probabilities to file

prob_name::String 

What to name the probability variable to file

Ronin.predict_with_model — Method

Simple function that opens a given h5 file with feature data and applies a specific model to it. Returns a tuple of predictions, targets. Also contains the ability to write these predictions and solutions out to a separate h5 file.

Required arguments

model_path::String 

Location of trained RF model (saved in jld2 file format)

input_h5::String 

Location of h5 file containing input features.

Optional Keyword Arguments

write_out::Bool = false

Whether or not to write the results out to a file

outfile::String = Path to write results to if write_out == true 

Results will be written in the h5 format with the name "Predicitions" and "Ground Truth"

probability_threshold::Float32 = .5

Minimum fraction of trees classifying a gate as meteorological for it to be assigned a label of meteorological

row_subset=:

Selection of rows to predict upon

Ronin.evaluate_model — Method

Function that takes in a given model, directory containing either cfradials or an already processed h5 file, a path to the configuration file, and a mode type ("C" for cfradials "H" for h5) and returns a Julia DataFrame containing a variety of metrics about the model's performance on the specified set, including precision and recall scores.

Required arguments

model_path::String

Path to input trained random forest model

input_file_dir::String

Path to input h5 file or directory of cfradial files to be processed

config_file_path 

Path to configuration file containing information about what features to calculate

Optional Arguments

mode::String = "C"

Whether to process a directory of cfradial files ("C" mode) or simply utilize an already-processed h5 file ("H" mode)

write_out::Bool = false

If in "C" mode, whether or not to write the resulting calculated features out to a file

output_file::String = "_.h5" 

Location to write calculated output features to.

col_subset = : 

Subset of columns of X array to input to model.

Also contains all keyword arguments for calculate_features

proba_seq::StepRangeLen = .1:.1:.9

Sequence of probabilities to iterate through when evaluating the model

Ronin.error_characteristics — Method

error_characteristics(file_path::String, config_file_path::String, model_path::String;
indexer_var::String="VV", QC_variable::String="VG", decision_threshold::Float64 = .5, write_out::Bool=false,
output_name::String="Model_Error_Characteristics.h5")

Function to process a set of cfradial files that have already been interactively QC'ed and return information about where errors occur in the files relative to model predictions. Requires a pre-trained model and configuration, as well as scans that have already been interactively quality controlled.

#Required Arguments

file_path::String

Path to file or directory of cfradials to be processed

config_file_path::String

Path to configuration file containing parameters to calculate for the cfradials

model_path::String

Path to pre-trained random forest model

Optional keyword arguments

indexer_var::String="VV" 

Name of a raw variable in input NetCDF files. Used to determine where missing data exists in the input sweeps. Data at these locations will be removed from the outputted features.

QC_variable::String="VG"

Name of variable in CFRadial files that has already been interactively QC'ed. Used as the verification data.

decision_threshold::Float64 = .5

Fraction of decision trees in the RF model that must agree for a given gate to be classified as meteorological. For example, at .5, >=50% of the trees must predict that a gate is meteorological for it to be classified as such, otherwise it is assigned as non-meteorological.

write_out::Bool=false

Whether or not to output the model evaluation data to an HDF5 file

output_name::String="Model_Error_Characteristics.h5"

Name/Path of desired HDF5 output location

Returns

Returns a tuple of (X, Y, indexer, predictions, falsepositives, falsenegatives) Where

X::Matrix{Float64}

Each row in X represents a different radar gate, while each column a different parameter as according to the order that they are listed in the configfilepath

Y::Matrix{Int64} 

Each row in Y represents a radar gate, and its classification according to the interactive QC applied to it.

indexer::Matrix{Int64}

For all gates in the input directory, contains 1 if the gate passed basic QC thresholds (Low NCP, etc.) and 0 if it did not. Useful if one wishes to reconstruct 2D scan from flattened data

predictions:Matrix{Int32}

Trained machine learning model predictions for the classification of a gate - 1 if predicted to be meteorological data, 0 otherwise.

false_postivies::BitMatrix

Which gates were misclassified as meteorological data relative to interactive QC

false_negatives::BitMatrix

Which gates were misclassified as non-meteorological data relative to interactive QC

Predicting using a composite model

Ronin.ModelConfig — Type

Stuct used to store configuration information for a given model

Required arguments

num_models:::Int64

Number of ML models in the model chain. Can be one or more.

model_output_paths::Vector{String}

Vector containing paths to each model in the model chain. Should be same length as the number of models

met_probs::Vector{Tuple{Float64,Float64}}

Vector containing the decision range for a gate to be considered meteorological in each model in the chain. Example, if set to (.9, 1), > 90% of trees in the random forest must assign a gate a label of meteorological for it to be considered meteorological. The range is exclusive on the lower end and inclusive on the high end. Form is (lowthreshold, highthreshold)

feature_output_paths::Vector{String}

Vector containing paths representing the locations to output calculated features to for each model in the chain.

input_path::String

Directory containing input radar data

input_config::String 

Path to file containing feature variables to calculate

file_preprocessed::Vector{Bool}

For each model in the chain, contains a boolean value signifying if the correspondant feature output path has already been processed. If true, will open the file at this path instead of re-calculating input features.

Optional arguments

verbose::Bool = true 

Whether to print out timing information, etc.

REMOVE_LOW_NCP::Bool = true

Whether to automatically remove gates that do not meet a basic NCP threshold

REMOVE_HIGH_PGG::Bool = true

Whether to automatically remove gates that do not meet a basic PGG threshold

HAS_INTERACTIVE_QC::Bool = false 

Whether the radar data has already had interactive QC applied to it

QC_var::String = "VG"

If radar data has interactive QC already applied, the name of a variable that the QC has been applied to

remove_var::String = "VV" 

Name of a raw variable in the radar data that can be used to determine the location of missing gates

replace_missing::Bool = false 

For spatial feature (AVG, STD, etc.) calculation, whether or not to replace MISSING gates in the mask area with FILL_VAL

write_out::Bool = true 

Whether or not to write the calculated input features to disk, paths specified in featureoutputpaths

QC_mask::Bool = false 

For the first model in the chain, whether or not to mask gates considered for feature calculation using a mask specified by mask_name More details elsewhere in the documentation.

mask_name::String = ""

See above

VARS_TO_QC::Vector{String} = ["VV", "ZZ"]

List of variables to apply QC to to get mask for next model in chain

QC_SUFFIX::String

Postfix to apply to variable name once QC has been applied.

class_weights::String = ""

Class weighting scheme to apply in the training of RF model. Currently only "balanced" is implemented.

n_trees::Int = 21

Number of trees in the random forest

max_depth::Int = 14

Maximum depth of any one tree in the random forest

Ronin.train_multi_model — Function

train_multi_model(config::ModelConfig)

All-in-one function to take in a set of radar data, calculate input features, and train a chain of random forest models for meteorological/non-meteorological gate identification.

#Required arguments

config::ModelConfig

Struct containing configuration info for model training

Ronin.composite_prediction — Function

composite_prediction(config::ModelConfig)

Passes feature data through a model or series of models and returns model classifications. Applies configuration such as masking and basic QC (high PGG/low NCP) specified by config

Returns

predictions::Vector{Bool} Model classifications for gates that passed basic quality control thresholds

values::BitVector Verification gates correspondant to predictions

init_idxers::Vector{Vector{Float64}} Information about where original radar data did/did not meet basic quality control thresholds. Each vector contains a flattened vector describing whether or not a given gate was predicted on.

total_met_probs::Vector{Float64}If kewyword argument returnprobs is set to true, then `totalmet_probs` will be returned. Each entry into this vector corresponds to the gate represented by predictions and values, and denotes the fraction of trees in the random forest that classified the gate as meteorological.

All values returned will be only those that passed quality control checks in the first pass of the model minimum NCP / PGG thresholds. In order to reconstruct a scan, user would need to use the values in the returned indexers.

Ronin.multipass_uncertain — Function

multipass_uncertain(base_config::ModelConfig, low_threshold::Float64, high_threshold::Float64; 
low_model_path::String = "low_config_model.jld2", skip_training_base=false)

Function that takes in a set of model parameters, trains a model on the entire dataset, and then isolates gates that trees in the forest have trouble agreeing upon. More specifically, will isolate gates that have > low_threshold but < high_threshold fraction of trees in agreement that they are meteorological. It will subsequently train a random forest model specifically on these gates, saving it to ambig_model.jld2

Required Arguments

base_config::ModelConfig Model configuration object that contains information about input data and where to write models to

low_threshold::Float64 Used for selecting ambiguous gates, the minimum fraction of trees that must agree for a datapoint to be considered

high_threshold::Float64 Used for selecting ambiguous gates, the maximum fraction of trees that may classify a gate as meteorological for it to be considered

Optional Arguments

low_model_path::String="low_config_model.jld2"` Where to output the base model (trained on all gates) to

ambig_model_path::String="ambig_model.jld2" Where to output model trained on difficult gates to

skip_training_base::Bool=false If base model has already been trained, set to true to skip retraining

mode::String = "train" Options are "train" or "predict". If mode == "predict", will not do any model training

Returns

Will return 3 items: A vector of predictions, a vector of verification data, and an indexer vector which can be used to determine which gates fall between the lowthreshold and highthreshold in the basic model. The prediction vector is a combination of classifications from the model trained on the base set and the model trained on the difficult gates.

Non-user facing

Ronin.get_task_params — Function

Function to parse a given task list Also performs checks to ensure that the specified tasks are able to be performed to the specified CFRad file

Parses input parameter file for use in outputting feature names to HDF5 file as attributes. NOTE: Cfradial-unaware. If one of the variables is specified incorrectly in the parameter file, will cause errors

Ronin.process_single_file — Function

Driver function that calculates a set of features from a single CFRadial file. Features are specified in file located at argfile_path.

Will return a tuple of (X, Y, indexer) where X is the features matrix, Y, a matrix containing the verification

where human QC determined the gate was meteorological (value of 1), or non-meteorological (value of 0),

and indexer contains a vector of booleans describing which gates met basic quality control thresholds and thus are represented in the X and Y matrixes

Weight matrixes are specified in file header, or passed as explicit argument.

Required arguments

cfrad::NCDataset 

Input NCDataset containing radar scan variables

argfile_path::String 

Path to file contaning config/task information.

Optional keyword arguments

HAS_INTERACTIVE_QC::Bool = false

If the scan has already had a human apply quality control to it, set to true. Otherwise, false

REMOVE_LOW_NCP::Bool = false

Whether or not to ignore gates that do not meet a minimum NCP/SQI threshold. If true, these gates will be set to false in indexer, and features/verification will not be calculated for them.

REMOVE_HIGH_PGG::Bool = false

Whether or not to ignore gates that exceed a given Probability of Ground Gate(PGG) threshold. If true, these gates will be set to false in indexer, and features/verification will not be calculated for them.

QC_variable::String = "VG"

Name of a variable in input CFRadial file that has had QC applied to it already. Used to calculate verification Y matrix.

remove_variable::String = "VV" 

Name of raw variable in input CFRadial file that will be used to determine where missing gates exist in the sweep.

replace_missing::Bool = false

For spatial parameters, whether or not to replace missings values with FILL_VAL

Returns:

-X: Array that is dimensioned (num_gates x num_features) where num_gates is the number of valid 
    (non-missing, meeting NCP/PGG thresholds) the function finds, and num_features is the 
    number of features specified in the argument file to calculate. 

-Y: IF HAS_INTERACTIVE_QC == true, will return Y, array containing 1 if a datapoint was retained 
    during interactive QC, and 0 otherwise. 

-INDEXER: Based on remove_variable as described above, contains boolean array specifiying
            where in the scan features valid data and where does not.