Basic Workflow Walkthrough

RadarQC is a package that utilizes a methodology developed by Dr. Alex DesRosiers and Dr. Michael Bell to remove non-meteorological gates from Doppler radar scans by leveraging machine learning techniques. In its current form, it contains functionality to derive a set of features from input radar data, use these features to train a Random Forest classification model, and apply this model to the raw fields contained within the radar scans. It also has some model evaluation ability. The beginning of this guide will walk through a basic workflow to train a model starting from scratch.

API / Function References

Calculating Model Input Features

Ronin.calculate_features — Method

Function to process a set of cfradial files and produce input features for training/evaluating a model

Required arguments

input_loc::String

Path to input cfradial or directory of input cfradials

argument_file::String

Path to configuration file containing which features to calculate

output_file::String

Path to output calculated features to (generally ends in .h5)

HAS_MANUAL_QC::Bool

Specifies whether or not the file(s) have already undergone a manual QC procedure. If true, function will also output a Y array used to verify where manual QC removed gates. This array is formed by considering where gates with non-missing data in raw scans (specified by remove_variable) are set to missing after QC is performed.

Optional keyword arguments

verbose::Bool

If true, will print out timing information as each file is processed

REMOVE_LOW_NCP::Bool

If true, will ignore gates with Normalized Coherent Power/Signal Quality Index below a threshold specified in RQCFeatures.jl

REMOVE_HIGH_PGG::Bool

If true, will ignore gates with Probability of Ground Gate (PGG) values at or above a threshold specified in RQCFeatures.jl

QC_variable::String

Name of variable in input NetCDF files that has been quality-controlled.

remove_variable::String

Name of a raw variable in input NetCDF files. Used to determine where missing data exists in the input sweeps. Data at these locations will be removed from the outputted features.

replace_missing 

Whether or not to replace MISSING values with FILL_VAL in spatial parameter calculations Default value: False

Ronin.calculate_features — Method

Function to process a set of cfradial files and produce input features for training/evaluating a model. Allows for user-specified tasks and weight matrices, otherwise the same as above.

Required arguments

input_loc::String

Path to input cfradial or directory of input cfradials

tasks::Vector{String}

Vector containing the features to be calculated for each cfradial. Example [DBZ, ISO(DBZ)]

weight_matrixes::Vector{Matrix{Union{Missing, Float64}}}

For each task, a weight matrix specifying how much each gate in a spatial calculation will be given. Required to be the same size as tasks

output_file::String 

Location to output the calculated feature data to.

HAS_MANUAL_QC::Bool

Specifies whether or not the file(s) have already undergone a manual QC procedure. If true, function will also output a Y array used to verify where manual QC removed gates. This array is formed by considering where gates with non-missing data in raw scans (specified by remove_variable) are set to missing after QC is performed.

Optional keyword arguments

verbose::Bool

If true, will print out timing information as each file is processed

REMOVE_LOW_NCP::Bool

If true, will ignore gates with Normalized Coherent Power/Signal Quality Index below a threshold specified in RQCFeatures.jl

REMOVE_HIGH_PGG::Bool

If true, will ignore gates with Probability of Ground Gate (PGG) values at or above a threshold specified in RQCFeatures.jl

QC_variable::String

Name of variable in input NetCDF files that has been quality-controlled.

remove_variable::String

Name of a raw variable in input NetCDF files. Used to determine where missing data exists in the input sweeps. Data at these locations will be removed from the outputted features.

replace_missing 

Whether or not to replace MISSING values with FILL_VAL in spatial parameter calculations Default value: False

Ronin.split_training_testing! — Function

Function to split a given directory or set of directories into training and testing files using the configuration described in DesRosiers and Bell 2023. This function assumes that input directories only contain cfradial files that follow standard naming conventions, and are thus implicitly chronologically ordered. The function operates by first dividing file names into training and testing sets following an 80/20 training/testing split, and subsequently softlinking each file to the training and testing directories. Attempts to avoid temporal autocorrelation while maximizing variance by dividing each case into several different training/testing sections.

Required Arguments:

DIR_PATHS::Vector{String}

List of directories containing cfradials to be used for model training/testing. Useful if input data is split into several different cases.

TRAINING_PATH::String 

Directory to softlink files designated for training into.

TESTING_PATH::String 

Directory to softlink files designated for testing into.

Ronin.train_model — Method

Function to train a random forest model using a precalculated set of input and output features (usually output from calculate_features). Returns nothing.

Required arguments

input_h5::String

Location of input features/targets. Input features are expected to have the name "X", and targets the name "Y". This should be taken care of automatically if they are outputs from calculate_features

model_location::String 

Path to save the trained model out to. Typically should end in .joblib

Optional keyword arguments

verify::Bool = false 

Whether or not to output a separate .h5 file containing the trained models predictions on the training set (Y_PREDICTED) as well as the targets for the training set (Y_ACTUAL)

verify_out::String="model_verification.h5"

If verify, the location to output this verification to.

Ronin.get_feature_importance — Method

Uses L1 regression with a variety of λ penalty values to determine the most useful features for

input to the random forest model.

Required Input

input_file_path::String 

Path to .h5 file containing model training features under ["X"] parameter, and model targets under ["Y"] parameter. Also expects the h5 file to contain an attribute known as Parameters containing abbreviations for the feature types

λs::Vector{Float64}

Vector of values used to vary the strength of the penalty term in the regularization.

Optional Keyword Arguments

pred_threshold::Float64

Minimum cofidence level for binary classifier when predicting

Returns

Returns a DataFrame with each row containing info about a regression for a specific λ, the values of the regression coefficients for each input feature, and the Root Mean Square Error of the resultant regression.

Applying and evaluating a trained model to data

Ronin.QC_scan — Function

Primary function to apply a trained RF model to certain raw fields of a cfradial scan. Values determined to be non-meteorological by the RF model will be replaced with Missing

Required Arguments

file_path::String 

Location of input cfradial or directory of cfradials one wishes to apply QC to

config_file_path::String 

Location of config file containing features to calculate as inputs to RF model

model_path::String 

Location of trained RF model (in joblib file format)

Optional Arguments

VARIABLES_TO_QC::Vector{String} = ["ZZ", "VV"]

List containing names of raw variables in the CFRadial to apply QC algorithm to.

QC_suffix::String = "_QC"

Used for naming the QC-ed variables in the modified CFRadial file. Field name will be QCsuffix appended to the raw field. Example: `DBZQC`

indexer_var::String = "VV"

Variable used to determine what gates are considered "missing" in the raw moments. QC will not be applied to these gates, they will simply remain missing.

decision_threshold::Float64 = .5

Used to leverage probablistic nature of random forest methodology. When the model has a greater than decision_threshold level confidence that a gate is meteorological data, it will be assigned as such. Anything at or below this confidence threshold will be assigned non-meteorological. At least in the ELDORA case, aggressive thresholds (.8 and above) have been found to maintain

92% of the meteorological data while removing >99% of non-meteorological gates.

Ronin.predict_with_model — Method

Simple function that opens a given h5 file with feature data and applies a specific model to it. Returns a tuple of predictions, targets.

Required arguments

model_path::String 

Location of trained RF model (saved in joblib file format)

input_h5::String 

Location of h5 file containing input features.

Ronin.evaluate_model — Method

Function that takes in a given model, directory containing either cfradials or an already processed h5 file, a path to the configuration file, and a mode type ("C" for cfradials "H" for h5) and returns a Julia DataFrame containing a variety of metrics about the model's performance on the specified set, including precision and recall scores.

Required arguments

model_path::String

Path to input trained random forest model

input_file_dir::String

Path to input h5 file or directory of cfradial files to be processed

config_file_path 

Path to configuration file containing information about what features to calculate

Optional Arguments

mode::String = "C"

Whether to process a directory of cfradial files ("C" mode) or simply utilize an already-processed h5 file ("H" mode)

write_out::Bool = false

If in "C" mode, whether or not to write the resulting calculated features out to a file

output_file::String = "_.h5" 

Location to write calculated output features to.

Also contains all keyword arguments for calculate_features

Non-user facing

Ronin.get_task_params — Function

Function to parse a given task list Also performs checks to ensure that the specified tasks are able to be performed to the specified CFRad file

Ronin.process_single_file — Function

Driver function that calculates a set of features from a single CFRadial file. Features are 
specified in file located at argfile_path. 

Will return a tuple of (X, Y, indexer) where X is the features matrix,  `Y`, a matrix containing the verification 
- where human QC determined the gate was meteorological (value of 1), or non-meteorological (value of 0), 
and `indexer` contains a vector of booleans describing which gates met basic quality control thresholds 
and thus are represented in the X and Y matrixes

Weight matrixes are specified in file header, or passed as explicit argument. 

# Required arguments 

```julia 
cfrad::NCDataset 
```
Input `NCDataset` containing radar scan variables 

```julia 
argfile_path::String 
```
Path to file contaning config/task information. 

# Optional keyword arguments 

```julia 
HAS_MANUAL_QC::Bool = false
```
If the scan has already had a human apply quality control to it, set to `true`. Otherwise, `false`

```julia 
REMOVE_LOW_NCP::Bool = false
```
Whether or not to ignore gates that do not meet a minimum NCP/SQI threshold. If `true`, these gates will be
set to `false` in `indexer`, and features/verification will not be calculated for them. 

```julia 
REMOVE_HIGH_PGG::Bool = false
```
Whether or not to ignore gates that exceed a given Probability of Ground Gate(PGG) threshold. If `true`, these gates will be
set to `false` in `indexer`, and features/verification will not be calculated for them. 

```julia 
QC_variable::String = "VG"
```
Name of a variable in input CFRadial file that has had QC applied to it already. Used to calculate verification `Y` matrix. 

```julia 
remove_variable::String = "VV" 
```
Name of raw variable in input CFRadial file that will be used to determine where `missing` gates exist in the sweep.

```julia 
replace_missing::Bool = false
```
For spatial parameters, whether or not to replace `missings` values with `FILL_VAL`

--- 

# Returns: 

    -X: Array that is dimensioned (num_gates x num_features) where num_gates is the number of valid 
        (non-missing, meeting NCP/PGG thresholds) the function finds, and num_features is the 
        number of features specified in the argument file to calculate. 

    -Y: IF HAS_MANUAL_QC == true, will return Y, array containing 1 if a datapoint was retained 
        during manual QC, and 0 otherwise. 

    -INDEXER: Based on remove_variable as described above, contains boolean array specifiying
              where in the scan features valid data and where does not.